add GasData data in form, dn refactor

+++ Models/HeatData.cs
@@ -8,7 +8,7 @@ namespace MediaWIiR_APP.Models
 {
     public class HeatData
     {
-        public int Heat { get; set; } //Ciepło i zmienna przesyl GJ
+        public decimal Heat { get; set; } //Ciepło i zmienna przesyl GJ
         public int Power { get; set; } // stala zmienna i zamowiona moc MW
         public int Month { get; set; } //Mc
         public int Carrier { get; set; } // Nośnik m3
